Fix CR/LF handling
Because the codec encodes every known character to a corresponding byte
value and every unknown character to a space, all newlines and carriage
returns were converted to spaces, since the character ROM does not
contain control characters.

While we could fit control characters in the unused regions of the A00
ROM, the A02 ROM is full. Instead, we change the `encode` function to
return an integer list instead of a byte array. Now negative integers
denote special characters, e.g. \n and \r.